[gnss] Configure for the PwrPak7-E1 in South Norway
This configuration files has two parts:
It connects to the Trondheim EUREF NTRIP caster. Username and password have been removed from the commit.
They can be obtained from www.euref-ip.be
It initializes the driver for a PwrPak7-E1 with parameters slightly different than the ones provided by Apollo for the Propak6.